+++ src/lib/constants.ts
@@ -5,46 +5,55 @@ export const mockupData: ITranslation[] = [
      {
           keyName: "greeting",
           baseString: "Hi",
-          translationString: "Բարև"
+          translationString: "Բարև",
+          lineNumber: 2
      },
      {
           keyName: "whatsUp",
           baseString: "What's Up?",
-          translationString: "Ինչպե՞ս ես"
+          translationString: "Ինչպե՞ս ես",
+          lineNumber: 3
      },
      {
           keyName: "bye",
           baseString: "Bye",
-          translationString: "Ցտեսություն"
+          translationString: "Ցտեսություն",
+          lineNumber: 4
      },
      {
           keyName: "yes",
           baseString: "Yes",
-          translationString: "Այո"
+          translationString: "Այո",
+          lineNumber: 5
      },
      {
           keyName: "no",
           baseString: "No",
-          translationString: "Ոչ"
+          translationString: "Ոչ",
+          lineNumber: 6
      },
      {
           keyName: "test",
           baseString: "Test",
-          translationString: "Test"
+          translationString: "Test",
+          lineNumber: 7
      },
      {
           keyName: "no",
           baseString: "No",
-          translationString: ""
+          translationString: "",
+          lineNumber: 8
      },
      {
           keyName: "common.buttons.save",
           baseString: "Save",
-          translationString: "Պահպանել"
+          translationString: "Պահպանել",
+          lineNumber: 11
      },
      {
           keyName: "common.buttons.cancel",
           baseString: "Cancel",
-          translationString: "Չեղարկել"
+          translationString: "Չեղարկել",
+          lineNumber: 12
      },
 ]