+++ src/lib/types.ts
@@ -2,6 +2,7 @@ export interface ITranslation{
      keyName: string,
      baseString: string,
      translationString: string,
+     lineNumber: number
 }
 
 export interface TreeNode {